changed problems with size_t

+++ mashkin.andrew/I2/maxCountOfEqualElements.cpp
@@ -1,7 +1,7 @@
 #include "maxCountOfEqualElements.h"
 #include <cstddef>
 
-void countMaxQuantityOfEqualElements(const int* arr, const int size, size_t& count, size_t& maxCount)
+void countMaxQuantityOfEqualElements(const int* arr, const size_t size, size_t& count, size_t& maxCount)
 {
   for (int i = 0; i < size - 1; ++i)
   {
