+++ mashkin.andrew/I2/samePositiveElements.h
@@ -1,5 +1,5 @@
 #ifndef SAMEPOSITIVEELEMENTS_H
 #define SAMEPOSITIVEELEMENTS_H
 
-bool isThereTheSamePositiveElements(const int* arr, const int size, bool& ans);
+bool isThereTheSamePositiveElements(const int* arr, const size_t size, bool& ans);
 #endif
